% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorMAP-package.R
\docType{package}
\name{FactorMAP-package}
\alias{FactorMAP}
\alias{FactorMAP-package}
\title{FactorMAP: Estimation and inference for factor models in Asset Pricing}
\description{
Estimation and inference for factor models in Asset Pricing.
}
\author{
\strong{Maintainer}: Alberto Quaini \email{alberto91quaini@gmail.com} (\href{https://orcid.org/0000-0002-1251-0599}{ORCID})

Authors:
\itemize{
  \item Ali Moin \email{amoin2k@gmail.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})
}

Other contributors:
\itemize{
  \item Gustavo Freire \email{gustavobcpf@hotmail.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID}) [contributor]
}

}
\keyword{internal}
