% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{factors5FF}
\alias{factors5FF}
\title{Fama-French 5 factors}
\format{
\subsection{\code{factors5FF}}{

A data frame with 719 rows and 6 columns:
\describe{
\item{Date}{Date in yyyymm format}
\item{...}{risk factor}
\item{...}{...}
}
}
}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
factors5FF
}
\description{
Monthly (percentage) excess returns on 5 Fama-French risk factors
from 07/1963 to 05/2023:
}
\keyword{datasets}
