% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_krsfrp.R
\name{IterativeKRSFRP}
\alias{IterativeKRSFRP}
\title{Perform the iterative model selection procedure for KRS factor risk premia
of Gospodinov Kan Robotti (2014) from moments extracted from data}
\usage{
IterativeKRSFRP(returns, factors, alpha = 0.05)
}
\arguments{
\item{returns}{\verb{n_observations x n_returns}-dimensional matrix of test asset
excess returns.}

\item{factors}{\verb{n_observations x n_factors}-dimensional matrix of risk
factors.}

\item{alpha}{significance level.}
}
\value{
list of original indices in "factors" of kept factors.
}
\description{
Performs the iterative model selection procedure for KRS factor
risk premia of Gospodinov Kan Robotti (2014) based on moments extracted from
factors and test asset excess returns.
}
\examples{
# import package data on 15 risk factors and 42 test asset excess returns
factors = FactorMAP::factors5FF[,-1]
returns = FactorMAP::returns25MEBTM[,-1]

# compute KRS factor risk premia and their standard errors
frp = FRP(returns, factors, include_standard_errors = TRUE)

}
