% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returnsIND17}
\alias{returnsIND17}
\title{Industry test Asset excess Returns}
\format{
\subsection{\code{returnsIND17}}{

A data frame with 719 rows and 18 columns:
\describe{
\item{Date}{Date in yyyymm format}
\item{...}{Asset excess returns}
\item{...}{...}
}
}
}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
returnsIND17
}
\description{
Monthly (percentage) excess returns on 17 industry portfolios
from 07/1963 to 05/2023:
}
\keyword{datasets}
