% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returnsMEBTM25}
\alias{returnsMEBTM25}
\title{Size/Book-to-market test Asset excess Returns}
\format{
\subsection{\code{returnsMEBTM25}}{

A data frame with 719 rows and 26 columns:
\describe{
\item{Date}{Date in yyyymm format}
\item{...}{Asset excess returns}
\item{...}{...}
}
}
}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
returnsMEBTM25
}
\description{
Monthly (percentage) excess returns on 25 portfolios sorted on
size and book-to-market from 07/1963 to 05/2023:
}
\keyword{datasets}
